Crate massa_network_exports

Manages a connection with a node

Re-exports

pub use settings::NetworkSettings;

Modules

settings

Macros

throw_handshake_error

Structs

BootstrapPeers

ConnectionCount

Connection count for a category

ConnectionId

A unique connection id for a node

NetworkCommandSender

NetworkEventReceiver

NetworkManager

Peer

PeerInfo

All information concerning a peer is here

Peers

Enums

ConnectionClosureReason

HandshakeErrorType

NetworkCommand

Commands that the worker can execute

NetworkConnectionErrorType

Incoming and outgoing connection with other peers error list

NetworkError

NetworkEvent

NetworkManagementCommand

PeerType

Peer categories. There is a defined number af slots for each category. Order matters: less prioritized peer type first

Type Definitions

Establisher

Listener

ReadHalf

WriteHalf