Module massa_network_exports::commands

Declaration of the public and internal Events and Commands used by massa-network that allow external communication with other nodes.

Operations work flow

An operation batch can be send or received by the massa node. Which operation is asked is managed by the NetworkWorker.

Other modules has the access to all commands but the usage if they want to send operation that they just noticed, they should use the command [NetworkCommand::SendOperationAnnouncements].

OurNode      ProtocolWorker      Network & NodeWorker
   |               |                |
   +------------------------------------------- Creation of some operations
   #               |                |           Extends the pool and
   #               |                |
   +-------------->|                |           Forward to Protocol
   |               +--------------->|           Forward to Network, then Node
   |               |                #
   |               |                #           Propagate the batch of Vec<OperationId> through the network

When receiving an operation batch from the network, the NodeWorker will inform the ProtocolWorker that some potentials new operations are in transit and can be requested.

The network will inform the node that new operations can transit with the propagation method SendOperationAnnouncements that we defined just before. Then, the node will manage if he ask or not the operations inside the ProtocolWorker on node-worker event ReceivedOperationAnnouncements

Asking for operations
---

NodeWorker      NetworkWorker          ProtocolWorker
   |               |                         |
   +------------------------------------------------------- Receive a batch of annoucemnt
   .               |                         |
   +-------------->|                         |              NetworkWorker react on previous event. Forward to protocol.
   |               +------------------------>#              - Check in the protocol if we already have operations
   |               |                         #              or not. Build the vector of requirement.
   |               |                         #              - Update the `NodeInfo` of the sender.
   |               |                         #              - Use the propagation algorithm
   |               |<------------------------+
   |<--------------+                         |
   |               |                         |
   |               |                         |
   |               |                         |              > Ask to the node that sent the batch a list of operations
   |               |                         |              > that we don't already know with `NodeCommand::AskForOperations`.
   |               |                         |
   +-------------->|                         |
   |               +------------------------>#
   |               |                         #
   |               |                         #              > Receive the full operations inside the structure
   |               |                         #              > `AskedOperation`.
   |               |                         #
   |               |                         #              Update local state and the `NodeInfo` of the sender if required.
   |               |                         |
   |               |<------------------------+              Once we received and store the operation, we can propagate it.
   |<--------------+                         |
   |               |                         |
   |               |                         |              > Propagate the batch through the network (send to nodes
   |               |                         |              > that don't already know the local operations (we suppose that
   |               |                         |              > from previous discussions with distant node)

Look at massa-protocol-worker/src/node-info.rs to look further how we remember which node know what.

Structs

NodeEvent

Events node worker can emit. Events are a tuple linking a node id to an event type

Enums

AskForBlocksInfo

Ask for the info about a block.

BlockInfoReply

A node replied with info about a block.

NetworkCommand

Commands that the worker can execute

NetworkEvent

network event

NetworkManagementCommand

Network management command

NodeCommand

network command

NodeEventType

Event types that node worker can emit Append on receive something from inside and outside. Outside initialization with Received prefix.