Enum massa_network_exports::NetworkError

#[non_exhaustive]
pub enum NetworkError {

    WrongSignature,
    GeneralProtocolError(String),
    ChannelError(String),
    TokioTaskJoinError(JoinError),
    TokieRecvError(RecvError),
    PeerConnectionError(NetworkConnectionErrorType),
    InvalidIpError(IpAddr),
    ActiveConnectionMissing(ConnectionId),
    IOError(Error),
    SerdeError(Error),
    MassaHashError(MassaHashError),
    HandshakeError(HandshakeErrorType),
    UnexpectedNodeCommandChannelClosure,
    UnexpectedWriterClosure,
    TimeError(TimeError),
    MissingPeersError,
    ModelsError(ModelsError),
    ContainerInconsistencyError(String),

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

WrongSignature

wrong signature

GeneralProtocolError(String)

Protocol err:{0}

ChannelError(String)

An error occurred during channel communication: {0}

TokioTaskJoinError(JoinError)

A tokio task has crashed err:{0}

TokieRecvError(RecvError)

error receiving oneshot response : {0}

PeerConnectionError(NetworkConnectionErrorType)

Error during network connection: {0}

InvalidIpError(IpAddr)

The ip:{0} address is not valid

ActiveConnectionMissing(ConnectionId)

Active connection missing:{0}

IOError(Error)

IO error : {0}

SerdeError(Error)

Serde error : {0}

MassaHashError(MassaHashError)

massa_hash error {0}

HandshakeError(HandshakeErrorType)

handshake error:{0:?}

UnexpectedNodeCommandChannelClosure

the network controller should not drop a node command sender before shutting down the node.

UnexpectedWriterClosure

the writer of a node should not drop its event sender before sending a clean_exit message.

TimeError(TimeError)

Time error {0}

MissingPeersError

missing peers

ModelsError(ModelsError)

models error: {0}

ContainerInconsistencyError(String)

container inconsistency error: {0}