Enum massa_network_exports::NodeEventType

pub enum NodeEventType {
    AskedPeerList,
    ReceivedPeerList(Vec<IpAddr>),
    ReceivedBlockHeader(WrappedHeader),
    ReceivedAskForBlocks(Vec<(BlockId, AskForBlocksInfo)>),
    ReceivedReplyForBlocks(Vec<(BlockId, BlockInfoReply)>),
    ReceivedOperations(Vec<WrappedOperation>),
    ReceivedOperationAnnouncements(OperationPrefixIds),
    ReceivedAskForOperations(OperationPrefixIds),
    ReceivedEndorsements(Vec<WrappedEndorsement>),
}

Event types that node worker can emit Append on receive something from inside and outside. Outside initialization with Received prefix.

Variants

AskedPeerList

Node we are connected to asked for advertised peers

ReceivedPeerList(Vec<IpAddr>)

Node we are connected to sent peer list

ReceivedBlockHeader(WrappedHeader)

Node we are connected to sent block header

ReceivedAskForBlocks(Vec<(BlockId, AskForBlocksInfo)>)

Node we are connected asked for info on a list of blocks.

ReceivedReplyForBlocks(Vec<(BlockId, BlockInfoReply)>)

Node we are connected sent info on a list of blocks.

ReceivedOperations(Vec<WrappedOperation>)

Received full operations.

ReceivedOperationAnnouncements(OperationPrefixIds)

Received an operation id batch announcing new operations

ReceivedAskForOperations(OperationPrefixIds)

Receive a list of wanted operations

ReceivedEndorsements(Vec<WrappedEndorsement>)

Receive a set of endorsement