Module massa_network_exports::network_controller

Modules

__mock_MockNetworkCommandSender

__mock_MockNetworkCommandSenderTrait

__mock_MockNetworkCommandSenderTrait_NetworkCommandSenderTrait

__mock_MockNetworkCommandSender_Clone

__mock_MockNetworkCommandSender_NetworkCommandSenderTrait

Structs

MockNetworkCommandSender

MockNetworkCommandSenderTrait

Network command sender interface. Can be mocked for testing

MockNetworkCommandSenderTrait_NetworkCommandSenderTrait 🔒

MockNetworkCommandSender_Clone 🔒

MockNetworkCommandSender_NetworkCommandSenderTrait 🔒

NetworkCommandSender

Network command sender

NetworkEventReceiver

network event receiver

NetworkManager

Network manager

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

META 🔒

META 🔒

META 🔒

Traits

NetworkCommandSenderTrait

Network command sender interface. Can be mocked for testing

Functions

make_runtime

Used by the bootstrap server to run async tasks, allowing the bootstrap module to remove the tokio dependency.

Opaque Types

🔒

🔒

🔒