Struct massa_network_exports::network_controller::MockNetworkCommandSender

pub struct MockNetworkCommandSender {
    Clone_expectations: MockNetworkCommandSender_Clone,
    NetworkCommandSenderTrait_expectations: MockNetworkCommandSender_NetworkCommandSenderTrait,
}

Fields
Clone_expectations: MockNetworkCommandSender_CloneNetworkCommandSenderTrait_expectations: MockNetworkCommandSender_NetworkCommandSenderTrait
Implementations

impl MockNetworkCommandSender

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockNetworkCommandSender

pub fn expect_clone(&mut self) -> &mut Expectation

Create an Expectation for mocking the clone method

impl MockNetworkCommandSender

pub fn expect_node_ban_by_ids(&mut self) -> &mut Expectation

Create an Expectation for mocking the node_ban_by_ids method

pub fn expect_node_ban_by_ips(&mut self) -> &mut Expectation

Create an Expectation for mocking the node_ban_by_ips method

pub fn expect_add_to_whitelist(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_to_whitelist method

pub fn expect_remove_from_whitelist(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_from_whitelist method

pub fn expect_node_unban_by_ids(&mut self) -> &mut Expectation

Create an Expectation for mocking the node_unban_by_ids method

pub fn expect_node_unban_ips(&mut self) -> &mut Expectation

Create an Expectation for mocking the node_unban_ips method

pub fn expect_send_block_info(&mut self) -> &mut Expectation

Create an Expectation for mocking the send_block_info method

pub fn expect_ask_for_block_list(&mut self) -> &mut Expectation

Create an Expectation for mocking the ask_for_block_list method

pub fn expect_send_block_header(&mut self) -> &mut Expectation

Create an Expectation for mocking the send_block_header method

pub fn expect_get_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_peers method

pub fn expect_get_network_stats(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_network_stats method

pub fn expect_get_bootstrap_peers(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_bootstrap_peers method

pub fn expect_send_operations(&mut self) -> &mut Expectation

Create an Expectation for mocking the send_operations method

pub fn expect_announce_operations(&mut self) -> &mut Expectation

Create an Expectation for mocking the announce_operations method

pub fn expect_send_ask_for_operations(&mut self) -> &mut Expectation

Create an Expectation for mocking the send_ask_for_operations method

pub fn expect_send_endorsements(&mut self) -> &mut Expectation

Create an Expectation for mocking the send_endorsements method

pub fn expect_node_sign_message(&mut self) -> &mut Expectation

Create an Expectation for mocking the node_sign_message method