Trait massa_network_exports::network_controller::NetworkCommandSenderTrait

pub trait NetworkCommandSenderTrait: Send + 'static {

    // Required methods
    fn node_ban_by_ids<'life0, 'async_trait>(
        &'life0 self,
        ids: Vec<NodeId>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn node_ban_by_ips<'life0, 'async_trait>(
        &'life0 self,
        ips: Vec<IpAddr>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_to_whitelist<'life0, 'async_trait>(
        &'life0 self,
        ips: Vec<IpAddr>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_from_whitelist<'life0, 'async_trait>(
        &'life0 self,
        ips: Vec<IpAddr>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn node_unban_by_ids<'life0, 'async_trait>(
        &'life0 self,
        ids: Vec<NodeId>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn node_unban_ips<'life0, 'async_trait>(
        &'life0 self,
        ips: Vec<IpAddr>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_block_info<'life0, 'async_trait>(
        &'life0 self,
        node: NodeId,
        info: Vec<(BlockId, BlockInfoReply)>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ask_for_block_list<'life0, 'async_trait>(
        &'life0 self,
        list: HashMap<NodeId, Vec<(BlockId, AskForBlocksInfo)>>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_block_header<'life0, 'async_trait>(
        &'life0 self,
        node: NodeId,
        header: SecuredHeader
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Peers, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_network_stats<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<NetworkStats, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_bootstrap_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<BootstrapPeers, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_operations<'life0, 'async_trait>(
        &'life0 self,
        node: NodeId,
        operations: Vec<SecureShareOperation>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn announce_operations<'life0, 'async_trait>(
        &'life0 self,
        to_node: NodeId,
        batch: OperationPrefixIds
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_ask_for_operations<'life0, 'async_trait>(
        &'life0 self,
        to_node: NodeId,
        wishlist: OperationPrefixIds
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_endorsements<'life0, 'async_trait>(
        &'life0 self,
        node: NodeId,
        endorsements: Vec<SecureShareEndorsement>
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn node_sign_message<'life0, 'async_trait>(
        &'life0 self,
        msg: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<PubkeySig, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Network command sender interface. Can be mocked for testing

Required Methods

fn node_ban_by_ids<'life0, 'async_trait>( &'life0 self, ids: Vec<NodeId> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

ban node(s) by id(s)

fn node_ban_by_ips<'life0, 'async_trait>( &'life0 self, ips: Vec<IpAddr> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

ban node(s) by ip(s)

fn add_to_whitelist<'life0, 'async_trait>( &'life0 self, ips: Vec<IpAddr> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

add ip to whitelist

fn remove_from_whitelist<'life0, 'async_trait>( &'life0 self, ips: Vec<IpAddr> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

remove ip from whitelist

fn node_unban_by_ids<'life0, 'async_trait>( &'life0 self, ids: Vec<NodeId> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

remove from banned node(s) by id(s)

fn node_unban_ips<'life0, 'async_trait>( &'life0 self, ips: Vec<IpAddr> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

remove from banned node(s) by ip(s)

fn send_block_info<'life0, 'async_trait>( &'life0 self, node: NodeId, info: Vec<(BlockId, BlockInfoReply)> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send info about the contents of a block.

fn ask_for_block_list<'life0, 'async_trait>( &'life0 self, list: HashMap<NodeId, Vec<(BlockId, AskForBlocksInfo)>> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send the order to ask for a block.

fn send_block_header<'life0, 'async_trait>( &'life0 self, node: NodeId, header: SecuredHeader ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send the order to send block header.

Note: with the current use of shared storage, sending a header requires having the block stored. This matches the current use of send_block_header, which is only used after a block has been integrated in the graph.

fn get_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Peers, NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send the order to get peers.

fn get_network_stats<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<NetworkStats, NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

get network stats

fn get_bootstrap_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<BootstrapPeers, NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send the order to get bootstrap peers.

fn send_operations<'life0, 'async_trait>( &'life0 self, node: NodeId, operations: Vec<SecureShareOperation> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

send operations to node

fn announce_operations<'life0, 'async_trait>( &'life0 self, to_node: NodeId, batch: OperationPrefixIds ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a new call to the network, sending a announcement of operation ID prefixes to a target node (to_node)

Returns

Can return a [NetworkError::ChannelError] that must be managed by the direct caller of the function.

fn send_ask_for_operations<'life0, 'async_trait>( &'life0 self, to_node: NodeId, wishlist: OperationPrefixIds ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a new call to the network, sending a wishlist of operationIds to a target node (to_node) in order to receive the full operations in the future.

Returns

Can return a [NetworkError::ChannelError] that must be managed by the direct caller of the function.

fn send_endorsements<'life0, 'async_trait>( &'life0 self, node: NodeId, endorsements: Vec<SecureShareEndorsement> ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

send endorsements to node id

fn node_sign_message<'life0, 'async_trait>( &'life0 self, msg: Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<PubkeySig, NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sign a message using the node’s keypair