Struct massa_network_exports::settings::NetworkConfig

pub struct NetworkConfig {

    pub bind: SocketAddr,
    pub routable_ip: Option<IpAddr>,
    pub protocol_port: u16,
    pub connect_timeout: MassaTime,
    pub wakeup_interval: MassaTime,
    pub initial_peers_file: PathBuf,
    pub peers_file: PathBuf,
    pub keypair_file: PathBuf,
    pub peer_types_config: EnumMap<PeerType, PeerTypeConnectionConfig>,
    pub max_in_connections_per_ip: usize,
    pub max_idle_peers: usize,
    pub max_banned_peers: usize,
    pub peers_file_dump_interval: MassaTime,
    pub message_timeout: MassaTime,
    pub ask_peer_list_interval: MassaTime,
    pub max_send_wait_node_event: MassaTime,
    pub max_send_wait_network_event: MassaTime,
    pub ban_timeout: MassaTime,
    pub peer_list_send_timeout: MassaTime,
    pub max_in_connection_overflow: usize,
    pub max_operations_per_message: u32,
    pub max_bytes_read: f64,
    pub max_bytes_write: f64,
    pub max_ask_blocks: u32,
    pub max_operations_per_block: u32,
    pub thread_count: u8,
    pub endorsement_count: u32,
    pub max_peer_advertise_length: u32,
    pub max_endorsements_per_message: u32,
    pub max_message_size: u32,
    pub max_datastore_value_length: u64,
    pub max_op_datastore_entry_count: u64,
    pub max_op_datastore_key_length: u8,
    pub max_op_datastore_value_length: u64,
    pub max_function_name_length: u16,
    pub max_parameters_size: u32,
    pub controller_channel_size: usize,
    pub event_channel_size: usize,
    pub node_command_channel_size: usize,
    pub node_event_channel_size: usize,
    pub last_start_period: u64,

}

Network configuration

Fields
bind: SocketAddr

Where to listen for communications.

routable_ip: Option<IpAddr>

Our own IP if it is routable, else None.

protocol_port: u16

Protocol port

connect_timeout: MassaTime

Time interval spent waiting for a response from a peer. In milliseconds

wakeup_interval: MassaTime

Network_worker will try to connect to available peers every wakeup_interval. In milliseconds

initial_peers_file: PathBuf

Path to the file containing initial peers.

peers_file: PathBuf

Path to the file containing known peers.

keypair_file: PathBuf

Path to the file containing our keypair

peer_types_config: EnumMap<PeerType, PeerTypeConnectionConfig>

Configuration for PeerType connections

max_in_connections_per_ip: usize

Limit on the number of in connections per ip.

max_idle_peers: usize

Limit on the number of idle peers we remember.

max_banned_peers: usize

Limit on the number of banned peers we remember.

peers_file_dump_interval: MassaTime

Peer database is dumped every peers_file_dump_interval in milliseconds

message_timeout: MassaTime

After message_timeout milliseconds we are no longer waiting on handshake message

ask_peer_list_interval: MassaTime

Every ask_peer_list_interval in milliseconds we ask every one for its advertisable peers list.

max_send_wait_node_event: MassaTime

Max wait time for sending a Node event.

max_send_wait_network_event: MassaTime

Max wait time for sending a Network event.

ban_timeout: MassaTime

Time after which we forget a node

peer_list_send_timeout: MassaTime

Timeout Duration when we send a PeerList in handshake

max_in_connection_overflow: usize

Max number of in connection overflowed managed by the handshake that send a list of peers

max_operations_per_message: u32

Max operations per message in the network to avoid sending to big data packet.

max_bytes_read: f64

Read limitation for a connection in bytes per seconds

max_bytes_write: f64

Write limitation for a connection in bytes per seconds

max_ask_blocks: u32

Max number ids in ask blocks message

max_operations_per_block: u32

Max operations per block

thread_count: u8

Thread count

endorsement_count: u32

Endorsement count

max_peer_advertise_length: u32

Max peer advertise length

max_endorsements_per_message: u32

Max endorsements per message

max_message_size: u32

Max message size

max_datastore_value_length: u64

Maximum length of a datastore value

max_op_datastore_entry_count: u64

Maximum entry in an operation datastore

max_op_datastore_key_length: u8

Maximum length of an operation datastore key

max_op_datastore_value_length: u64

Maximum length of an operation datastore value

max_function_name_length: u16

Maximum length function name in call SC

max_parameters_size: u32

Maximum size of parameters in call SC

controller_channel_size: usize

Controller channel size

event_channel_size: usize

Event channel size

node_command_channel_size: usize

Node command channel size

node_event_channel_size: usize

Node event channel size

last_start_period: u64

last start period, used in message deserialization

Implementations

impl NetworkConfig

pub fn scenarios_default(port: u16, peers_file: &Path) -> Self

default network settings from port and peer file path