Crate massa_network_exports

Manages a connection with a node

Re-exports

pub use settings::NetworkConfig;

Modules

settings

network settings

Macros

throw_handshake_error

return handshake error

Structs

BootstrapPeers

Peers that are transmitted during bootstrap

BootstrapPeersDeserializer

Deserializer for BootstrapPeers

BootstrapPeersSerializer

Serializer for BootstrapPeers

ConnectionCount

Connection count for a category

ConnectionId

A unique connection id for a node

NetworkCommandSender

Network command sender

NetworkEventReceiver

network event receiver

NetworkManager

Network manager

NodeEvent

Events node worker can emit. Events are a tuple linking a node id to an event type

Peer

Associate a peer info with nodes

PeerInfo

All information concerning a peer is here

Peers

peers

Enums

AskForBlocksInfo

Ask for the info about a block.

BlockInfoReply

A node replied with info about a block.

ConnectionClosureReason

Possible reasons for a connection closure

HandshakeErrorType

Handshake error type

NetworkCommand

Commands that the worker can execute

NetworkConnectionErrorType

Incoming and outgoing connection with other peers error list

NetworkError

Network error

NetworkEvent

network event

NetworkManagementCommand

Network management command

NodeCommand

network command

NodeEventType

Event types that node worker can emit Append on receive something from inside and outside. Outside initialization with Received prefix.

PeerType

Peer categories. There is a defined number of slots for each category. Order matters: less prioritized peer type first

Type Definitions

Establisher

Connection establisher

Listener

Connection listener

ReadHalf

read half of the TCP stream

WriteHalf

Write half of the TCP stream